// backend/src/modules/achievements/enums/achievements-codes.ts
export enum AchievementsCodes {
    WELCOME = 'WELCOME',
    ROOKIE = 'ROOKIE',
    ENDURANCE_TRAINING = 'ENDURANCE_TRAINING',
    CONSISTENT = 'CONSISTENT',
    BESTIE = 'BESTIE',
    VIRTUOSO = 'VIRTUOSO',
    ADVENTUROUS = 'ADVENTUROUS',
    RELIABLE = 'RELIABLE',
    EARLY_RISER = 'EARLY_RISER',
    RELAXED = 'RELAXED',
    FANATIC = 'FANATIC',
    SEX_GIANT = 'SEX_GIANT',
    CREATIVE = 'CREATIVE',
    EXPLORER = 'EXPLORER',
    IMPRESSIVE_STAMINA = 'IMPRESSIVE_STAMINA',
}